# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ootils Core API",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/ingest/calendars": {
      "post": {
        "tags": [
          "calendars"
        ],
        "summary": "Import calendar",
        "description": "Import non-working days for a location. Upsert per (location × date). Missing entry = working day by default.",
        "operationId": "ingest_calendars_v1_ingest_calendars_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestCalendarsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestCalendarsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/calendars/{location_external_id}": {
      "get": {
        "tags": [
          "calendars"
        ],
        "summary": "Get calendar",
        "description": "List calendar entries for a location (optional: filter by date range).",
        "operationId": "get_calendars_v1_calendars__location_external_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "location_external_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location External Id"
            }
          },
          {
            "name": "from_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Date"
            }
          },
          {
            "name": "to_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Date"
            }
          },
          {
            "name": "working_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Working Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCalendarsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/calendars/working-days": {
      "post": {
        "tags": [
          "calendars"
        ],
        "summary": "Compute working days",
        "description": "Return the date resulting from adding N working days to a start date, respecting the location calendar.",
        "operationId": "compute_working_days_v1_calendars_working_days_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkingDaysRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkingDaysResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CalendarEntryInput": {
        "properties": {
          "calendar_date": {
            "type": "string",
            "format": "date",
            "title": "Calendar Date",
            "description": "Calendar date (YYYY-MM-DD)."
          },
          "is_working_day": {
            "type": "boolean",
            "title": "Is Working Day",
            "description": "False = non-working day (holiday, closure). Default = False.",
            "default": false
          },
          "shift_count": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 3.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Shift Count",
            "description": "Number of shifts on this day [0–3]. Optional."
          },
          "capacity_factor": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 2.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity Factor",
            "description": "Capacity factor [0.0–2.0]. 1.0 = normal capacity. Optional."
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Notes (e.g. 'Christmas', 'Planned maintenance'). Optional."
          }
        },
        "type": "object",
        "required": [
          "calendar_date"
        ],
        "title": "CalendarEntryInput"
      },
      "CalendarEntryOutput": {
        "properties": {
          "calendar_date": {
            "type": "string",
            "format": "date",
            "title": "Calendar Date"
          },
          "is_working_day": {
            "type": "boolean",
            "title": "Is Working Day"
          },
          "shift_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shift Count"
          },
          "capacity_factor": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity Factor"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "calendar_date",
          "is_working_day",
          "shift_count",
          "capacity_factor",
          "notes"
        ],
        "title": "CalendarEntryOutput"
      },
      "CalendarIngestSummary": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "inserted": {
            "type": "integer",
            "title": "Inserted"
          },
          "updated": {
            "type": "integer",
            "title": "Updated"
          },
          "errors": {
            "type": "integer",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "total",
          "inserted",
          "updated",
          "errors"
        ],
        "title": "CalendarIngestSummary"
      },
...
      "GetCalendarsResponse": {
        "properties": {
          "location_external_id": {
            "type": "string",
            "title": "Location External Id"
          },
          "location_id": {
            "type": "string",
            "title": "Location Id"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/CalendarEntryOutput"
            },
            "type": "array",
            "title": "Entries"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "location_external_id",
          "location_id",
          "entries",
          "total"
        ],
        "title": "GetCalendarsResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestCalendarsRequest": {
        "properties": {
          "location_external_id": {
            "type": "string",
            "title": "Location External Id",
            "description": "External_id of the target location."
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/CalendarEntryInput"
            },
            "type": "array",
            "title": "Entries",
            "description": "Calendar entries to import (upsert per location × date)."
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "description": "If true, validation only — no DB writes.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "location_external_id",
          "entries"
        ],
        "title": "IngestCalendarsRequest"
      },
      "IngestCalendarsResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "location_external_id": {
            "type": "string",
            "title": "Location External Id"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "summary": {
            "$ref": "#/components/schemas/CalendarIngestSummary"
          }
        },
        "type": "object",
        "required": [
          "status",
          "location_external_id",
          "location_id",
          "summary"
        ],
        "title": "IngestCalendarsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkingDaysRequest": {
        "properties": {
          "location_external_id": {
            "type": "string",
            "title": "Location External Id",
            "description": "Location external_id."
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date",
            "description": "Start date (YYYY-MM-DD)."
          },
          "add_working_days": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Add Working Days",
            "description": "Number of working days to add (>= 0)."
          }
        },
        "type": "object",
        "required": [
          "location_external_id",
          "start_date",
          "add_working_days"
        ],
        "title": "WorkingDaysRequest"
      },
      "WorkingDaysResponse": {
        "properties": {
          "location_external_id": {
            "type": "string",
            "title": "Location External Id"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "add_working_days": {
            "type": "integer",
            "title": "Add Working Days"
          },
          "result_date": {
            "type": "string",
            "format": "date",
            "title": "Result Date"
          },
          "working_days_found": {
            "type": "integer",
            "title": "Working Days Found"
          },
          "non_working_days_skipped": {
            "type": "integer",
            "title": "Non Working Days Skipped"
          }
        },
        "type": "object",
        "required": [
          "location_external_id",
          "start_date",
          "add_working_days",
          "result_date",
          "working_days_found",
          "non_working_days_skipped"
        ],
        "title": "WorkingDaysResponse"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}